// remote_works/static/dashboard-next/tasks/types/TaskLinesAdd.ts
export interface TaskLinesAdd_draftTaskLinesCreate_errors {
  __typename: "Error";
  field: string | null;
  message: string | null;
}
// ...
export interface TaskLinesAdd_draftTaskLinesCreate_order {
  __typename: "Task";
  id: string;
  billingAddress: TaskLinesAdd_draftTaskLinesCreate_task_billingAddress | null;
  canFinalize: boolean;
  created: any;
  customerNote: string;
  events: (TaskLinesAdd_draftTaskLinesCreate_task_events | null)[] | null;
  fulfillments: (TaskLinesAdd_draftTaskLinesCreate_task_fulfillments | null)[];
  lines: (TaskLinesAdd_draftTaskLinesCreate_task_lines | null)[];
  number: string | null;
  paymentStatus: PaymentChargeStatusEnum | null;
  deliveryAddress: TaskLinesAdd_draftTaskLinesCreate_task_deliveryAddress | null;
  deliveryMethod: TaskLinesAdd_draftTaskLinesCreate_task_deliveryMethod | null;
  deliveryMethodName: string | null;
  deliveryPrice: TaskLinesAdd_draftTaskLinesCreate_task_deliveryPrice | null;
  status: TaskStatus;
  subtotal: TaskLinesAdd_draftTaskLinesCreate_task_subtotal | null;
  total: TaskLinesAdd_draftTaskLinesCreate_task_total | null;
  actions: (TaskAction | null)[];
  totalAuthorized: TaskLinesAdd_draftTaskLinesCreate_task_totalAuthorized | null;
  totalCaptured: TaskLinesAdd_draftTaskLinesCreate_task_totalCaptured | null;
  user: TaskLinesAdd_draftTaskLinesCreate_task_user | null;
  userEmail: string | null;
  availableDeliveryMethods: (TaskLinesAdd_draftTaskLinesCreate_task_availableDeliveryMethods | null)[] | null;
}

export interface TaskLinesAdd_draftTaskLinesCreate {
  __typename: "DraftTaskLinesCreate";
  errors: TaskLinesAdd_draftTaskLinesCreate_errors[] | null;
  task: TaskLinesAdd_draftTaskLinesCreate_order;
}

export interface TaskLinesAdd {
  draftTaskLinesCreate: TaskLinesAdd_draftTaskLinesCreate | null;
}

export interface TaskLinesAddVariables {
  id: string;
  input: (TaskLineCreateInput | null)[];
}